perf(sidebar): let the palette fetch its own lists, so closed routes never register them
